@@ -10,21 +10,15 @@
   "author": "Vladimir Feskov",
   "license": "MIT",
   "devDependencies": {
-    "@types/chai": "^4.2.12",
-    "@types/mocha": "^7.0.2",
-    "@types/mongodb": "^3.5.25",
-    "@types/mongoose": "^5.7.36",
-    "@types/nock": "^11.1.0",
-    "@types/node-fetch": "^2.5.7",
-    "chai": "^4.2.0",
-    "mocha": "^7.2.0",
-    "nock": "^12.0.3",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7"
+    "@types/node-fetch": "^2.6.2",
+    "ts-node": "^10.9.1",
+    "type-fest": "^2.19.0",
+    "typescript": "^4.7.4"
   },
+  "type": "module",
   "dependencies": {
-    "mongodb": "^3.6.0",
-    "mongoose": "^5.9.28",
-    "node-fetch": "2.6.0"
+    "mongoose": "^6.5.2",
+    "node-fetch": "^3.2.10",
+    "timeout-signal": "^1.1.0"
   }
 }
