@@ -1,33 +1,27 @@
-import { MongoClient } from "mongodb";
-import loadUsers from "./parts/loadUsers";
-import groupByRepo from "./parts/groupByRepo";
-import filterOutIrrelevantRepos from "./parts/filterOutIrrelevantRepos";
-import fetchTags from "./parts/fetchTags";
-import backToUsers from "./parts/backToUsers";
-import findUsersToAlert from "./parts/findUsersToAlert";
-import fetchReleaseNotes from "./parts/fetchReleaseNotes";
-import sendEmailAndUpdateDb from "./parts/sendEmailAndUpdateDb";
-import log, { debug } from "gitpunch-lib/log";
-import * as githubAtom from "gitpunch-lib/githubAtom";
-const { MONGODB_URL } = process.env;
+import loadUsers from "./parts/loadUsers.js";
+import groupByRepo from "./parts/groupByRepo.js";
+import filterOutIrrelevantRepos from "./parts/filterOutIrrelevantRepos.js";
+import fetchTags from "./parts/fetchTags.js";
+import backToUsers from "./parts/backToUsers.js";
+import findUsersToAlert from "./parts/findUsersToAlert.js";
+import fetchReleaseNotes from "./parts/fetchReleaseNotes.js";
+import sendEmailAndUpdateDb from "./parts/sendEmailAndUpdateDb.js";
+import log, { debug } from "gitpunch-lib/log.js";
+import * as githubAtom from "gitpunch-lib/githubAtom.js";
+import { connect, disconnect } from "gitpunch-lib/db/index.js";
 
 export async function handler(event, context, callback) {
-  let client: MongoClient;
   githubAtom.trackTotalRequests();
   try {
-    client = await MongoClient.connect(MONGODB_URL, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    });
-    const collection = client.db().collection("users");
+    await connect();
     debug("loadUsers");
-    const dbUsers = await loadUsers(collection);
+    const dbUsers = await loadUsers();
     debug("groupByRepo");
     const byRepo = groupByRepo(dbUsers);
     debug("filterOutIrrelevantRepos");
-    const byRepoRelevant = await filterOutIrrelevantRepos(client, byRepo);
+    const byRepoRelevant = await filterOutIrrelevantRepos(byRepo);
     debug("fetchTags");
-    const byRepoWithTags = await fetchTags(client, byRepoRelevant);
+    const byRepoWithTags = await fetchTags(byRepoRelevant);
     debug("backToUsers");
     const fullUsers = backToUsers(byRepoWithTags);
     debug("findUsersToAlert");
@@ -35,11 +29,11 @@ export async function handler(event, context, callback) {
     debug("fetchReleaseNotes");
     const withReleaseNotes = await fetchReleaseNotes(usersToAlert);
     debug("sendEmailAndUpdateDb");
-    await sendEmailAndUpdateDb(withReleaseNotes, collection);
+    await sendEmailAndUpdateDb(withReleaseNotes);
   } catch (e) {
     log("error", { error: { message: e.message, stack: e.stack } });
   }
-  client && client.close();
+  await disconnect();
   githubAtom.closeConnections();
   githubAtom.logTotalRequests();
   callback();
