@@ -1,8 +1,8 @@
-import { Collection } from "mongodb";
-import log from "gitpunch-lib/log";
+import log from "gitpunch-lib/log.js";
+import { UserModel } from "gitpunch-lib/db/index.js";
 import { DBUser } from "./interfaces";
 
-export default async function loadUsers(collection: Collection) {
+export default async function loadUsers(): Promise<DBUser[]> {
   const query: any = {
     watching: true,
     $or: [
@@ -20,16 +20,22 @@ export default async function loadUsers(collection: Collection) {
     });
   }
 
-  let users: DBUser[] = [];
-  const cursor = collection.find(query);
-  while (await cursor.hasNext()) {
-    const user = await cursor.next();
-    user.alerted = (user.alerted || []).reduce((alerted, [repo, tag]) => {
-      alerted[repo] = tag;
-      return alerted;
-    }, {});
-    users.push(user);
-  }
+  const users = await UserModel.find(query).then((users) =>
+    users.map((user) => ({
+      _id: user._id,
+      email: user.email,
+      mutedRepos: user.mutedRepos,
+      accessToken: user.accessToken,
+      majors: user.majors,
+      minors: user.minors,
+      patches: user.patches,
+      repos: user.repos,
+      alerted: (user.alerted || []).reduce((alerted, [repo, tag]) => {
+        alerted[repo] = tag;
+        return alerted;
+      }, {}),
+    }))
+  );
   log("dbUsers", { count: users.length });
   return users;
 }
