@@ -1,22 +1,19 @@
-import log from "gitpunch-lib/log";
-import { Collection } from "mongodb";
+import { UserModel } from "gitpunch-lib/db/User.js";
+import log from "gitpunch-lib/log.js";
 import {
   ONLY_UPDATE_ALERTED,
   SEND_EMAIL_AND_UPDATE_ALERTED,
-} from "./constants";
-import Email from "./email";
+} from "./constants.js";
+import Email from "./email.js";
 import { ActionableUser, Alerted } from "./interfaces";
 const { assign, keys } = Object;
 
-export default function sendEmailAndUpdateDb(
-  users: ActionableUser[],
-  collection: Collection
-) {
+export default function sendEmailAndUpdateDb(users: ActionableUser[]) {
   return Promise.all(
     users.map(async (user) => {
       try {
         await sendEmail(user);
-        await updateDb(user, collection);
+        await updateDb(user);
       } catch (error) {
         const { _id, email } = user;
         log("alertAndUpdateDbError", { _id, email, error: error.stack });
@@ -25,9 +22,9 @@ export default function sendEmailAndUpdateDb(
   );
 }
 
-const MAJOR = /^v?\d+\.0\.0$/;
-const MINOR = /^v?\d+\.\d+\.0$/;
-const PATCH = /^v?\d+\.\d+\.\d+$/;
+const MAJOR = /^v?\d+(\.0)?(\.0)?$/;
+const MINOR = /^v?\d+(\.\d+)?(\.0)?$/;
+const PATCH = /^v?\d+(\.\d+)?(\.\d+)?$/;
 export async function sendEmail(user: ActionableUser): Promise<any> {
   const {
     email,
@@ -61,10 +58,7 @@ function filteredTags(tags, pattern) {
   return tags.filter(({ name }) => pattern.test(name));
 }
 
-export async function updateDb(
-  user: ActionableUser,
-  collection: Collection
-): Promise<any> {
+export async function updateDb(user: ActionableUser): Promise<any> {
   const { _id, email, alerted, actionableRepos } = user;
   const $set: any = {};
   const repos = [
@@ -79,5 +73,5 @@ export async function updateDb(
   const alertedAsArray = keys(alerted).map((repo) => [repo, alerted[repo]]);
   $set.alerted = alertedAsArray;
   log("updateDb", { _id, email, $set });
-  return collection.updateOne({ _id }, { $set });
+  return UserModel.updateOne({ _id }, { $set });
 }
